adapted test to new structure
